feat: Post 관련 API 응답 시 좋아요, 생성일, 수정일 반환

+++ src/main/java/capssungzzang/_dayscottonball/domain/post/dto/PostResponse.java
@@ -12,7 +12,7 @@ public class PostResponse {
     private Long postId;
     private String title;
     private String content;
-    private Integer likes;
+    private long likes;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
